requirements.txt
groq==0.5.0
streamlit==1.29.0
pydantic==2.5.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-groq==0.1.0